tools/requirements.txt
pygit2==1.7.2
bidict==0.22.0
Pillow==9.3.0
PyYaml==6.0.1

# changelogs
beautifulsoup4==4.9.3

# ezdb
mysql-connector-python==8.0.33

# codecheck
ripgrepy==2.0.0
toposort==1.10
